+++ rodbus-ffi-new/src/channel.rs
@@ -1,11 +1,9 @@
-use rodbus::client::session::CallbackSession;
-use rodbus::types::{AddressRange, UnitId, WriteMultiple};
+use rodbus::types::{AddressRange, WriteMultiple};
 use std::ffi::CStr;
 use std::net::SocketAddr;
 use std::os::raw::c_char;
 use std::ptr::null_mut;
 use std::str::FromStr;
-use tokio::time::Duration;
 
 pub struct Channel {
     pub(crate) inner: rodbus::client::channel::Channel,
@@ -303,229 +301,4 @@ pub(crate) unsafe fn channel_write_multiple_registers_async(
     channel
         .runtime
         .block_on(session.write_multiple_registers(argument, callback));
-}
-
-impl crate::ffi::BitReadCallback {
-    pub(crate) fn bad_argument(self) {
-        let result = crate::ffi::BitReadResult {
-            result: crate::ffi::ErrorInfo::error(crate::ffi::Status::BadArgument),
-            iterator: null_mut(),
-        };
-
-        self.on_complete(result);
-    }
-
-    pub(crate) fn to_fn_once(
-        self,
-    ) -> impl FnOnce(std::result::Result<rodbus::types::BitIterator, rodbus::error::Error>) -> ()
-    {
-        move |result: std::result::Result<rodbus::types::BitIterator, rodbus::error::Error>| {
-            match result {
-                Err(err) => {
-                    self.on_complete(err.into());
-                }
-                Ok(values) => {
-                    let mut iter = crate::BitIterator::new(values);
-
-                    let result = crate::ffi::BitReadResult {
-                        result: crate::ffi::ErrorInfo::success(),
-                        iterator: &mut iter as *mut crate::BitIterator,
-                    };
-
-                    self.on_complete(result);
-                }
-            }
-        }
-    }
-}
-
-impl crate::ffi::RequestParam {
-    pub(crate) fn build_session(
-        &self,
-        channel: &Channel,
-    ) -> rodbus::client::session::CallbackSession {
-        CallbackSession::new(channel.inner.create_session(
-            UnitId::new(self.unit_id),
-            Duration::from_millis(self.timeout_ms as u64),
-        ))
-    }
-}
-
-impl crate::ffi::RegisterReadCallback {
-    pub(crate) fn bad_argument(self) {
-        let result = crate::ffi::RegisterReadResult {
-            result: crate::ffi::ErrorInfo::error(crate::ffi::Status::BadArgument),
-            iterator: null_mut(),
-        };
-        self.on_complete(result);
-    }
-
-    pub(crate) fn to_fn_once(
-        self,
-    ) -> impl FnOnce(std::result::Result<rodbus::types::RegisterIterator, rodbus::error::Error>) -> ()
-    {
-        move |result: std::result::Result<rodbus::types::RegisterIterator, rodbus::error::Error>| {
-            match result {
-                Err(err) => {
-                    self.on_complete(err.into());
-                }
-                Ok(values) => {
-                    let mut iter = crate::RegisterIterator::new(values);
-
-                    let result = crate::ffi::RegisterReadResult {
-                        result: crate::ffi::ErrorInfo::success(),
-                        iterator: &mut iter as *mut crate::RegisterIterator,
-                    };
-
-                    self.on_complete(result);
-                }
-            }
-        }
-    }
-}
-
-impl crate::ffi::ResultCallback {
-    pub(crate) fn bad_argument(self) {
-        let result = crate::ffi::ErrorInfo::error(crate::ffi::Status::BadArgument);
-        self.on_complete(result);
-    }
-
-    /// we do't care what type T is b/c we're going to ignore it
-    pub(crate) fn to_fn_once<T>(
-        self,
-    ) -> impl FnOnce(std::result::Result<T, rodbus::error::Error>) -> () {
-        move |result: std::result::Result<T, rodbus::error::Error>| match result {
-            Err(err) => {
-                self.on_complete(err.into());
-            }
-            Ok(_) => {
-                self.on_complete(crate::ffi::ErrorInfo::success());
-            }
-        }
-    }
-}
-
-impl crate::ffi::ErrorInfo {
-    pub(crate) fn error(err: crate::ffi::Status) -> Self {
-        Self {
-            summary: err,
-            exception: crate::ffi::Exception::Unknown,
-            raw_exception: 0,
-        }
-    }
-
-    pub(crate) fn success() -> Self {
-        Self {
-            summary: crate::ffi::Status::Ok,
-            exception: crate::ffi::Exception::Unknown,
-            raw_exception: 0,
-        }
-    }
-}
-
-impl<'a> std::convert::From<rodbus::error::Error> for crate::ffi::RegisterReadResult<'a> {
-    fn from(err: rodbus::error::Error) -> Self {
-        Self {
-            result: err.into(),
-            iterator: null_mut(),
-        }
-    }
-}
-
-impl<'a> std::convert::From<rodbus::error::Error> for crate::ffi::BitReadResult<'a> {
-    fn from(err: rodbus::error::Error) -> Self {
-        Self {
-            result: err.into(),
-            iterator: null_mut(),
-        }
-    }
-}
-
-impl From<rodbus::error::Error> for crate::ffi::ErrorInfo {
-    fn from(err: rodbus::error::Error) -> Self {
-        fn from_status(status: crate::ffi::Status) -> crate::ffi::ErrorInfo {
-            crate::ffi::ErrorInfo {
-                summary: status,
-                exception: crate::ffi::Exception::Unknown, // doesn't matter what it is
-                raw_exception: 0,
-            }
-        }
-
-        match err {
-            rodbus::error::Error::Internal(_) => from_status(crate::ffi::Status::InternalError),
-            rodbus::error::Error::NoConnection => from_status(crate::ffi::Status::NoConnection),
-            rodbus::error::Error::BadFrame(_) => from_status(crate::ffi::Status::BadFraming),
-            rodbus::error::Error::Shutdown => from_status(crate::ffi::Status::Shutdown),
-            rodbus::error::Error::ResponseTimeout => {
-                from_status(crate::ffi::Status::ResponseTimeout)
-            }
-            rodbus::error::Error::BadRequest(_) => from_status(crate::ffi::Status::BadRequest),
-            rodbus::error::Error::Exception(ex) => ex.into(),
-            rodbus::error::Error::Io(_) => from_status(crate::ffi::Status::IOError),
-            rodbus::error::Error::BadResponse(_) => from_status(crate::ffi::Status::BadResponse),
-        }
-    }
-}
-
-impl<'a> From<rodbus::error::details::ExceptionCode> for crate::ffi::ErrorInfo {
-    fn from(x: rodbus::error::details::ExceptionCode) -> Self {
-        fn from_exception(
-            exception: crate::ffi::Exception,
-            raw_exception: u8,
-        ) -> crate::ffi::ErrorInfo {
-            crate::ffi::ErrorInfo {
-                summary: crate::ffi::Status::Exception,
-                exception,
-                raw_exception,
-            }
-        }
-
-        match x {
-            rodbus::error::details::ExceptionCode::Acknowledge => {
-                from_exception(crate::ffi::Exception::Acknowledge, x.into())
-            }
-            rodbus::error::details::ExceptionCode::GatewayPathUnavailable => {
-                from_exception(crate::ffi::Exception::GatewayPathUnavailable, x.into())
-            }
-            rodbus::error::details::ExceptionCode::GatewayTargetDeviceFailedToRespond => {
-                from_exception(
-                    crate::ffi::Exception::GatewayTargetDeviceFailedToRespond,
-                    x.into(),
-                )
-            }
-            rodbus::error::details::ExceptionCode::IllegalDataAddress => {
-                from_exception(crate::ffi::Exception::IllegalDataAddress, x.into())
-            }
-            rodbus::error::details::ExceptionCode::IllegalDataValue => {
-                from_exception(crate::ffi::Exception::IllegalDataValue, x.into())
-            }
-            rodbus::error::details::ExceptionCode::IllegalFunction => {
-                from_exception(crate::ffi::Exception::IllegalFunction, x.into())
-            }
-            rodbus::error::details::ExceptionCode::MemoryParityError => {
-                from_exception(crate::ffi::Exception::MemoryParityError, x.into())
-            }
-            rodbus::error::details::ExceptionCode::ServerDeviceBusy => {
-                from_exception(crate::ffi::Exception::ServerDeviceBusy, x.into())
-            }
-            rodbus::error::details::ExceptionCode::ServerDeviceFailure => {
-                from_exception(crate::ffi::Exception::ServerDeviceFailure, x.into())
-            }
-            rodbus::error::details::ExceptionCode::Unknown(x) => {
-                from_exception(crate::ffi::Exception::Unknown, x)
-            }
-        }
-    }
-}
-
-impl std::convert::From<crate::ffi::Bit> for rodbus::types::Indexed<bool> {
-    fn from(x: crate::ffi::Bit) -> Self {
-        rodbus::types::Indexed::new(x.index, x.value)
-    }
-}
-
-impl std::convert::From<crate::ffi::Register> for rodbus::types::Indexed<u16> {
-    fn from(x: crate::ffi::Register) -> Self {
-        rodbus::types::Indexed::new(x.index, x.value)
-    }
 }
