+++ rodbus-ffi-new/src/lib.rs
@@ -1,4 +1,8 @@
 mod channel;
+// From<T> implementations for FFI types
+mod conversions;
+// Additional impl for FFI types
+mod ext;
 mod iterator;
 mod list;
 mod logging;
