+++ rodbus/examples/server.rs
@@ -105,7 +105,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     // print log messages to the console
-    SimpleLogger::new().with_level(log::LevelFilter::Info).init().unwrap();
+    SimpleLogger::new()
+        .with_level(log::LevelFilter::Info)
+        .init()
+        .unwrap();
 
     let handler =
         SimpleHandler::new(vec![false; 10], vec![false; 20], vec![0; 10], vec![0; 20]).wrap();
